conditional requests for actionRoute and projectRoute

+++ actionRouter.js
@@ -5,7 +5,7 @@ const db = require("./data/helpers/actionModel.js");
 router.get("/", async (req, res) => {
     try {
         const actions = await db.get();
-        actions.status(200).json(actions);
+        res.status(200).json(actions);
     } catch(error){
         res.status(500).json({ message: "error"})
     }
@@ -14,7 +14,7 @@ router.get("/", async (req, res) => {
 router.get("/:id", async (req, res) => {
     const { id } = req.params;
     try{
-        const action = db.get(id);
+        const action = await db.get(id);
         if(action){
             res.status(200).json(action);
         } else {
@@ -28,7 +28,12 @@ router.get("/:id", async (req, res) => {
 router.post("/", async (req, res) => {
     const action = req.body;
     try {
-        db.insert(action);
+        const newAction = await db.insert(action);
+        if (newAction) {
+            res.status(200).json(newAction);
+        }else {
+            res.status(500).json({ message: "error"})
+        }
     } catch(error){
         res.status(500).json({ message: "error"})
     }
@@ -37,7 +42,18 @@ router.post("/", async (req, res) => {
 router.delete("/:id", async (req, res) => {
     const { id } = req.params;
     try{
-        db.remove(id);
+        const action = await db.get(id);
+        if (action){
+            const deleted = await db.remove(id);
+            if(deleted) {
+                res.status(200).json(action);
+            } else {
+                res.status(500).json({ message: "could not delete"})
+            } 
+           
+        } else{
+            res.status(404).json({ message: "error"})
+        }
     } catch(error){
         res.status(500).json({ message: "error"});
     }
@@ -47,7 +63,12 @@ router.put("/:id", async (req, res) => {
     const { id } = req.params;
     const action = req.body;
     try {
-        db.update(id, action);
+        const edited = await db.update(id, action);
+        if (edited) {
+            res.status(200).json(edited);
+        } else {
+            res.status(404).json({ message: "error"});
+        }
     } catch(error) {
         res.status(500).json({ message: "error"})
     }
